+++ amd/src/view.js
@@ -22,12 +22,10 @@
  */
 
 import $ from 'jquery';
-import {get_string as getString} from 'core/str';
 
 export const init = () => {
-    // getString('logintopingo', 'mod_pingo')
-    //     .then(buttonString => {
-    //         $('.path-mod-pingo .loginform #id_submitbutton').attr('value', buttonString);
-    //     })
-    //     .catch();
+    $("#id_tag").on('change', function() {
+        $('input[name=reload]').val(1);
+        this.form.submit();
+    });
 };