+++ classes/privacy/provider.php
@@ -15,7 +15,7 @@
 // along with Moodle.  If not, see <https://www.gnu.org/licenses/>.
 
 /**
- * Privacy subsystem implementation for pingo.
+ * Privacy subsystem implementation for the PINGO plugin.
  *
  * @package    mod_pingo
  * @copyright  2023 coactum GmbH
@@ -37,7 +37,7 @@ use \core_privacy\local\request\user_preference_provider;
 
 
 /**
- * Implementation of the privacy subsystem plugin provider for the pingo activity module.
+ * Implementation of the privacy subsystem plugin provider for the PINGO activity module.
  *
  * @copyright  2023 coactum GmbH
  * @license    https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
@@ -62,23 +62,17 @@ class provider implements
      */
     public static function get_metadata(collection $items) : collection {
 
-        /* // The table 'pingo_participants' stores all pingo participants and their data.
-        $items->add_database_table('pingo_participants', [
-            'pingo' => 'privacy:metadata:pingo_participants:pingo',
-        ], 'privacy:metadata:pingo_participants');
+        // The table 'pingo_connections' stores all data for pingo connections.
+        $items->add_database_table('pingo_connections', [
+            'pingo' => 'privacy:metadata:pingo_connections:userid',
+            'userid' => 'privacy:metadata:pingo_connections:pingo',
+            'authentication_token' => 'privacy:metadata:pingo_connections:authentication_token',
+            'timestarted' => 'privacy:metadata:pingo_connections:timestarted',
+        ], 'privacy:metadata:pingo_connections');
 
-        // The table 'pingo_submissions' stores all group subbissions.
-        $items->add_database_table('pingo_submissions', [
-            'pingo' => 'privacy:metadata:pingo_submissions:pingo',
-        ], 'privacy:metadata:pingo_submissions');
+        // The plugin does not use any subsystems that save personal data.
 
-        // The plguin uses multiple subsystems that save personal data.
-        $items->add_subsystem_link('core_files', [], 'privacy:metadata:core_files');
-        $items->add_subsystem_link('core_rating', [], 'privacy:metadata:core_rating');
-        $items->add_subsystem_link('core_message', [], 'privacy:metadata:core_message');
-
-        // User preferences in the plugin.
-        $items->add_user_preference('pingo_sortoption', 'privacy:metadata:preference:pingo_sortoption'); */
+        // No user preferences in the plugin.
 
         return $items;
     }
@@ -86,7 +80,7 @@ class provider implements
     /**
      * Get the list of contexts that contain user information for the specified user.
      *
-     * In this case of all pingos where a user is exam participant.
+     * In this case of all PINGO instances where a user is authenticated to PINGO.
      *
      * @param   int         $userid     The user to search.
      * @return  contextlist $contextlist  The contextlist containing the list of contexts used in this plugin.
@@ -100,17 +94,15 @@ class provider implements
             'userid' => $userid,
         ];
 
-        // Get contexts of ... .
-
-        $sql;
-        /* $sql = "SELECT c.id
+        // Get contexts.
+        $sql = "SELECT c.id
                   FROM {context} c
                   JOIN {course_modules} cm ON cm.id = c.instanceid AND c.contextlevel = :contextlevel
                   JOIN {modules} m ON m.id = cm.module AND m.name = :modulename
-                  JOIN {pingo} e ON e.id = cm.instance
-                  JOIN {pingo_participants} p ON p.pingo = e.id
-                  WHERE p.moodleuserid = :userid
-        "; */
+                  JOIN {pingo} p ON p.id = cm.instance
+                  JOIN {pingo_connections} c ON c.pingo = p.id
+                  WHERE c.userid = :userid
+        ";
 
         $contextlist->add_from_sql($sql, $params);
 
@@ -136,12 +128,12 @@ class provider implements
 
         // Get users.
         $sql;
-        /* $sql = "SELECT p.moodleuserid
+        $sql = "SELECT c.userid
                   FROM {course_modules} cm
                   JOIN {modules} m ON m.id = cm.module AND m.name = :modulename
-                  JOIN {pingo} e ON e.id = cm.instance
-                  JOIN {pingo_participants} p ON p.pingo = e.id
-                 WHERE cm.id = :instanceid"; */
+                  JOIN {pingo} p ON p.id = cm.instance
+                  JOIN {pingo_connections} c ON c.pingo = p.id
+                 WHERE cm.id = :instanceid";
         $userlist->add_from_sql('userid', $sql, $params);
     }
 
@@ -163,7 +155,7 @@ class provider implements
         list($contextsql, $contextparams) = $DB->get_in_or_equal($contextlist->get_contextids(), SQL_PARAMS_NAMED);
         $params = $contextparams;
 
-        /* $sql = "SELECT
+        $sql = "SELECT
                 c.id AS contextid,
                 p.*,
                 cm.id AS cmid
@@ -192,177 +184,85 @@ class provider implements
                     // Write generic module intro files.
                     helper::export_context_files($context, $user);
 
-                    self::export_entries_data($userid, $pingo->id, $pingo->contextid);
-
-                    self::export_annotations_data($userid, $pingo->id, $pingo->contextid);
-
+                    self::export_connection_data($userid, $pingo->id, $pingo->contextid);
                 }
 
             }
         }
 
-        $pingos->close(); */
+        $pingos->close();
     }
 
     /**
-     * Store all information about all ....
+     * Store all information about all connections.
      *
      * @param   int         $userid The userid of the user whose data is to be exported.
      * @param   int         $pingoid The id of the pingo.
      * @param   int         $pingocontextid The context id of the pingo.
      */
-    /* protected static function export_entries_data(int $userid, $pingoid, $pingocontextid) {
+    protected static function export_connections_data(int $userid, $pingoid, $pingocontextid) {
         global $DB;
 
-        // Find all entries for this pingo written by the user.
+        // Find all connections for the PINGO instance.
         $sql = "SELECT
-                    e.id,
-                    e.pingo,
-                    e.userid,
-                    e.timecreated,
-                    e.timemodified,
-                    e.text,
-                    e.format,
-                    e.rating,
-                    e.feedback,
-                    e.formatfeedback,
-                    e.teacher,
-                    e.timemarked,
-                    e.baseentry
-                   FROM {pingo_entries} e
+                    c.id,
+                    c.userid,
+                    c.pingo,
+                    c.authentication_token,
+                    c.timestarted
+                   FROM {pingo_connections} c
                    WHERE (
-                    e.pingo = :pingoid AND
-                    e.userid = :userid
+                    c.pingo = :pingoid AND
+                    c.userid = :userid
                     )
         ";
 
         $params['userid'] = $userid;
         $params['pingoid'] = $pingoid;
 
-        // Get the pingos from the entries.
-        $entries = $DB->get_recordset_sql($sql, $params);
+        $connections = $DB->get_recordset_sql($sql, $params);
 
-        if ($entries->valid()) {
-            foreach ($entries as $entry) {
-                if ($entry) {
+        if ($connections->valid()) {
+            foreach ($connections as $connection) {
+                if ($connection) {
                     $context = \context::instance_by_id($pingocontextid);
 
-                    self::export_entry_data($userid, $context, ['pingo-entry-' . $entry->id], $entry);
+                    self::export_connection_data($userid, $context, ['pingo-connection-' . $connection->id], $connection);
                 }
             }
         }
 
-        $entries->close();
-    } */
+        $connections->close();
+    }
 
     /**
-     * Export all data in the entry.
+     * Export all data for the connection.
      *
      * @param   int         $userid The userid of the user whose data is to be exported.
-     * @param   \context    $context The instance of the pingo context.
+     * @param   \context    $context The instance of the PINGO context.
      * @param   array       $subcontext The location within the current context that this data belongs.
-     * @param   \stdClass   $entry The entry.
+     * @param   \stdClass   $connection The connection.
      */
-    /* protected static function export_entry_data(int $userid, \context $context, $subcontext, $entry) {
+    protected static function export_connection_data(int $userid, \context $context, $subcontext, $connection) {
 
-        if ($entry->timecreated != 0) {
-            $timecreated = transform::datetime($entry->timecreated);
+        if ($connection->timestarted != 0) {
+            $timestarted = transform::datetime($connection->timestarted);
         } else {
-            $timecreated = null;
-        }
-
-        if ($entry->timemodified != 0) {
-            $timemodified = transform::datetime($entry->timemodified);
-        } else {
-            $timemodified = null;
-        }
-
-        if ($entry->timemarked != 0) {
-            $timemarked = transform::datetime($entry->timemarked);
-        } else {
-            $timemarked = null;
+            $timestarted = null;
         }
 
         // Store related metadata.
-        $entrydata = (object) [
-            'pingo' => $entry->pingo,
-            'userid' => $entry->userid,
-            'timecreated' => $timecreated,
-            'timemodified' => $timemodified,
-            'rating' => $entry->rating,
-            'teacher' => $entry->teacher,
-            'timemarked' => $timemarked,
-            'baseentry' => $entry->baseentry,
+        $connectiondata = (object) [
+            'userid' => $connection->userid,
+            'pingo' => $connection->pingo,
+            'authentication_token' => $connection->authentication_token,
+            'timestarted' => $timestarted,
         ];
-
-        $entrydata->text = writer::with_context($context)->rewrite_pluginfile_urls($subcontext, 'mod_pingo',
-            'entry', $entry->id, $entry->text);
-
-        $entrydata->text = format_text($entrydata->text, $entry->format, (object) [
-            'para' => false,
-            'context' => $context,
-        ]);
-
-        $entrydata->feedback = writer::with_context($context)->rewrite_pluginfile_urls($subcontext, 'mod_pingo',
-            'feedback', $entry->id, $entry->feedback);
-
-        $entrydata->feedback = format_text($entrydata->feedback, $entry->formatfeedback, (object) [
-            'para' => false,
-            'context' => $context,
-        ]);
 
         // Store the entry data.
         writer::with_context($context)
-            ->export_data($subcontext, $entrydata)
-            ->export_area_files($subcontext, 'mod_pingo', 'entry', $entry->id)
-            ->export_area_files($subcontext, 'mod_pingo', 'feedback', $entry->id);
-
-        // Store all ratings against this entry as the entry belongs to the user. All ratings on it are ratings of their content.
-        \core_rating\privacy\provider::export_area_ratings($userid, $context, $subcontext, 'mod_pingo',
-            'entry', $entry->id, false);
-    } */
-
-    /**
-     * Store all user preferences for the plugin.
-     *
-     * @param   int         $userid The userid of the user whose data is to be exported.
-     */
-    /* public static function export_user_preferences(int $userid) {
-        $user = \core_user::get_user($userid);
-
-        if ($pingosortoption = get_user_preferences('pingo_sortoption', 0, $userid)) {
-            switch ($pingosortoption) {
-                case 1:
-                    $sortoption = get_string('currenttooldest', 'mod_pingo');
-                    break;
-                case 2:
-                    $sortoption = get_string('oldesttocurrent', 'mod_pingo');
-                    break;
-                case 3:
-                    $sortoption = get_string('lowestgradetohighest', 'mod_pingo');
-                    break;
-                case 4:
-                    $sortoption = get_string('highestgradetolowest', 'mod_pingo');
-                    break;
-                default:
-                    $sortoption = get_string('currenttooldest', 'mod_pingo');
-                    break;
-            }
-
-            writer::export_user_preference('mod_pingo', 'pingo_sortoption', $pingosortoption, $sortoption);
-        }
-
-        if ($pingopagecount = get_user_preferences('pingo_pagecount', 0, $userid)) {
-            writer::export_user_preference('mod_pingo', 'pingo_pagecount', $pingopagecount,
-                get_string('privacy:metadata:preference:pingo_pagecount', 'mod_pingo'));
-        }
-
-        if ($pingoactivepage = get_user_preferences('pingo_activepage', 0, $userid)) {
-            writer::export_user_preference('mod_pingo', 'pingo_activepage', $pingoactivepage,
-                get_string('privacy:metadata:preference:pingo_activepage', 'mod_pingo'));
-        }
-    } */
-
+            ->export_data($subcontext, $connectiondata);
+    }
 
     /**
      * Delete all data for all users in the specified context.
@@ -382,24 +282,10 @@ class provider implements
             return;
         }
 
-        // Delete advanced grading information (not implemented yet).
-
-        /* // Delete all ratings in the context.
-        \core_rating\privacy\provider::delete_ratings($context, 'mod_pingo', 'entry');
-
-        // Delete all files from the entry.
-        $fs = get_file_storage();
-        $fs->delete_area_files($context->id, 'mod_pingo', 'entry');
-        $fs->delete_area_files($context->id, 'mod_pingo', 'feedback');
-
         // Delete all records.
-        if ($DB->record_exists('pingo_participants', ['pingo' => $cm->instance])) {
-            $DB->delete_records('pingo_participants', ['pingo' => $cm->instance]);
+        if ($DB->record_exists('pingo_connections', ['pingo' => $cm->instance])) {
+            $DB->delete_records('pingo_connections', ['pingo' => $cm->instance]);
         }
-
-        if ($DB->record_exists('pingo_submissions', ['pingo' => $cm->instance])) {
-            $DB->delete_records('pingo_submissions', ['pingo' => $cm->instance]);
-        } */
     }
 
     /**
@@ -417,40 +303,14 @@ class provider implements
             // Get the course module.
             $cm = $DB->get_record('course_modules', ['id' => $context->instanceid]);
 
-            /* // Handle any advanced grading method data first (not implemented yet).
+            // Delete connections for user.
+            if ($DB->record_exists('pingo_connections', ['pingo' => $cm->instance, 'userid' => $userid])) {
 
-            // Delete ratings.
-            $entriessql = "SELECT
-                                e.id
-                                FROM {pingo_entries} e
-                                WHERE (
-                                    e.pingo = :pingoid AND
-                                    e.userid = :userid
-                                )
-            ";
-
-            $entriesparams = [
-                'pingoid' => $cm->instance,
-                'userid' => $userid,
-            ];
-
-            \core_rating\privacy\provider::delete_ratings_select($context, 'mod_pingo',
-                'entry', "IN ($entriessql)", $entriesparams);
-
-            // Delete all files from the entries.
-            $fs = get_file_storage();
-            $fs->delete_area_files_select($context->id, 'mod_pingo', 'entry', "IN ($entriessql)", $entriesparams);
-            $fs->delete_area_files_select($context->id, 'mod_pingo', 'feedback', "IN ($entriessql)", $entriesparams);
-
-            // Delete entries for user.
-            if ($DB->record_exists('pingo_entries', ['pingo' => $cm->instance, 'userid' => $userid])) {
-
-                $DB->delete_records('pingo_entries', [
+                $DB->delete_records('pingo_connections', [
                     'pingo' => $cm->instance,
                     'userid' => $userid,
                 ]);
-
-            } */
+            }
         }
     }
 
@@ -468,29 +328,9 @@ class provider implements
         list($userinsql, $userinparams) = $DB->get_in_or_equal($userlist->get_userids(), SQL_PARAMS_NAMED);
         $params = array_merge(['pingoid' => $cm->instance], $userinparams);
 
-        // Handle any advanced grading method data first (not implemented yet).
-
-        // Delete ratings.
-        /* $entriesselect = "SELECT
-                            e.id
-                            FROM {pingo_entries} e
-                            WHERE (
-                                e.pingo = :pingoid AND
-                                userid {$userinsql}
-                            )
-        ";
-
-        \core_rating\privacy\provider::delete_ratings_select($context, 'mod_pingo', 'entry', "IN ($entriesselect)", $params);
-
-        // Delete all files from the entries.
-        $fs = get_file_storage();
-        $fs->delete_area_files_select($context->id, 'mod_pingo', 'entry', "IN ($entriesselect)", $params);
-        $fs->delete_area_files_select($context->id, 'mod_pingo', 'feedback', "IN ($entriesselect)", $params);
-
-        // Delete entries for users.
-        if ($DB->record_exists_select('pingo_entries', "pingo = :pingoid AND userid {$userinsql}", $params)) {
-            $DB->delete_records_select('pingo_entries', "pingo = :pingoid AND userid {$userinsql}", $params);
-        } */
-
+        // Delete connections for users.
+        if ($DB->record_exists_select('pingo_connections', "pingo = :pingoid AND userid {$userinsql}", $params)) {
+            $DB->delete_records_select('pingo_connections', "pingo = :pingoid AND userid {$userinsql}", $params);
+        }
     }
 }