+++ lib.php
@@ -48,14 +48,6 @@ function pingo_supports($feature) {
             return true;
         case FEATURE_SHOW_DESCRIPTION:
             return true;
-        // case FEATURE_GRADE_HAS_GRADE:
-        //     return true;
-        // case FEATURE_RATE:
-        //     return true;
-        // case FEATURE_GROUPS:
-        //     return true;
-        // case FEATURE_GROUPINGS:
-        //     return true;
         case FEATURE_COMPLETION_TRACKS_VIEWS:
             return true;
         case FEATURE_BACKUP_MOODLE2: